+++ tests/web/feed.py
@@ -20,6 +20,14 @@ class FeedTests(WebApiTestBase):
                 'name': 'test_timeline_feed',
                 'test': FeedTests('test_timeline_feed', api),
             },
+            {
+                'name': 'test_reels_tray',
+                'test': FeedTests('test_reels_tray', api),
+            },
+            {
+                'name': 'test_reels_feed',
+                'test': FeedTests('test_reels_feed', api),
+            },
         ]
 
     def test_tag_feed(self):
@@ -43,3 +51,15 @@ class FeedTests(WebApiTestBase):
         self.assertIsNotNone(results.get('user', {}).get('username'))
         self.assertGreater(
             len(results.get('user', {}).get('edge_web_feed_timeline', {}).get('edges', [])), 0)
+
+    def test_reels_tray(self):
+        results = self.api.reels_tray().get('data', {})
+        self.assertGreater(
+            len(results.get('user', {}).get(
+                'feed_reels_tray', {}).get(
+                    'edge_reels_tray_to_reel', {}).get('edges', [])), 0)
+
+    def test_reels_feed(self):
+        results = self.api.reels_feed(['25025320']).get('data', {})
+        self.assertGreater(
+            len(results.get('reels_media')), 0)