+++ instagram_private_api/__init__.py
@@ -4,7 +4,9 @@ from .client import Client
 from .compatpatch import ClientCompatPatch
 from .errors import (
     ClientError, ClientLoginError, ClientLoginRequiredError,
-    ClientCookieExpiredError, ClientThrottledError, ClientConnectionError
+    ClientCookieExpiredError, ClientThrottledError, ClientConnectionError,
+    ClientCheckpointRequiredError, ClientChallengeRequiredError,
+    ClientSentryBlockError,
 )
 from .endpoints.upload import MediaRatios
 from .endpoints.common import MediaTypes