+++ instagram_private_api/endpoints/upload.py
@@ -8,7 +8,7 @@ from ..compat import (
     compat_urllib_error, compat_urllib_request,
     compat_http_client
 )
-from ..errors import ClientError, ClientConnectionError
+from ..errors import ErrorHandler, ClientError, ClientConnectionError
 from ..http import MultipartFormDataEncoder
 from ..utils import (
     max_chunk_count_generator, max_chunk_size_generator,
@@ -445,17 +445,9 @@ class UploadEndpointsMixin(object):
             self.logger.debug('POST {0!s}'.format(endpoint_url))
             response = self.opener.open(req, timeout=self.timeout)
         except compat_urllib_error.HTTPError as e:
-            error_msg = e.reason
             error_response = self._read_response(e)
             self.logger.debug('RESPONSE: {0:d} {1!s}'.format(e.code, error_response))
-            try:
-                error_obj = json.loads(error_response)
-                if error_obj.get('message'):
-                    error_msg = '{0!s}: {1!s}'.format(e.reason, error_obj['message'])
-            except ValueError as ve:
-                # do nothing else, prob can't parse json
-                self.logger.warn('Error parsing error response: {}'.format(str(ve)))
-            raise ClientError(error_msg, e.code, error_response)
+            ErrorHandler.process(e, error_response)
         except (SSLError, timeout, SocketError,
                 compat_urllib_error.URLError,   # URLError is base of HTTPError
                 compat_http_client.HTTPException) as connection_error:
@@ -630,17 +622,9 @@ class UploadEndpointsMixin(object):
                                         code=500)
 
                     except compat_urllib_error.HTTPError as e:
-                        error_msg = e.reason
                         error_response = self._read_response(e)
                         self.logger.debug('RESPONSE: {0:d} {1!s}'.format(e.code, error_response))
-                        try:
-                            error_obj = json.loads(error_response)
-                            if error_obj.get('message'):
-                                error_msg = '{0!s}: {1!s}'.format(e.reason, error_obj['message'])
-                        except ValueError as ve:
-                            # do nothing else, prob can't parse json
-                            self.logger.warn('Error parsing error response: {}'.format(str(ve)))
-                        raise ClientError(error_msg, e.code, error_response)
+                        ErrorHandler.process(e, error_response)
 
                     except (SSLError, timeout, SocketError,
                             compat_urllib_error.URLError,   # URLError is base of HTTPError